requirements.txt
requests~=2.28.2
python-dotenv==0.19.0
psycopg2-binary==2.9.5
requests
telegram~=0.0.1
Pyrogram~=2.0.30
python-telegram-bot==20.1
aiohttp~=3.8.4
aiogram~=2.25.1
pytz~=2022.7.1
aiopg~=1.4.0